Limit flash usage
Squashed commit of the following:

commit 467b3640959b1d1ec35c765701b0a520a86fed3a
Author: tx_haggis <13982343+adbancroft@users.noreply.github.com>
Date:   Sun Sep 22 22:30:42 2024 -0500

    Allow host to turn inlining off or on.

commit aa8d93eba9d3bd03f24d5fdd1848e5e2b338ead6
Author: tx_haggis <13982343+adbancroft@users.noreply.github.com>
Date:   Sat Sep 21 00:23:19 2024 -0500

    Limit defensive checks

commit 72e6d3ce79ea833da227735bf5f209a2b6b718f5
Author: tx_haggis <13982343+adbancroft@users.noreply.github.com>
Date:   Sat Sep 21 00:01:20 2024 -0500

    Place source in a cpp file, instead of inline: lowers the impact of loop unrolling. No performance penalty
