+++ test/test_performance/main.cpp
@@ -46,3 +46,5 @@ void loop()
     digitalWrite(LED_BUILTIN, LOW);
     delay(250);
 }
+
+#include "avr-fast-div.cpp"
