+++ test/test_performance/test_fast_div_performance.cpp
@@ -44,7 +44,7 @@ static void test_fast_div_perf_u16_u8_optimal(void)
   static constexpr index_range_generator<uint16_t> dividendGen = create_optimal_dividend_range<uint8_t, uint16_t>(divisorGen); 
 
 #if defined(UNOPTIMIZED_BUILD)  
-  constexpr uint8_t percentExpected = 80;
+  constexpr uint8_t percentExpected = 75;
 #else
   constexpr uint8_t percentExpected = 40;
 #endif 
@@ -58,7 +58,7 @@ static void test_fast_div_perf_u16_u8_worst_case(void)
   index_range_generator<uint16_t> dividendGen(UINT16_MAX/divisorGen.rangeMax()*3, UINT16_MAX, divisorGen.num_steps());
 
 #if defined(UNOPTIMIZED_BUILD)
-  constexpr uint8_t percentExpected = 115;
+  constexpr uint8_t percentExpected = 111;
 #else
   constexpr uint8_t percentExpected = 110;
 #endif 
@@ -73,7 +73,7 @@ static void test_fast_div_perf_u32_u8(void)
 #if defined(UNOPTIMIZED_BUILD)
   constexpr uint8_t percentExpected = 60;
 #else
-  constexpr uint8_t percentExpected = 40;
+  constexpr uint8_t percentExpected = 35;
 #endif 
   performance_test(16, dividendGen, divisorGen, percentExpected);
 }
@@ -87,7 +87,7 @@ static void test_fast_div_perf_u32_u16_optimal(void)
 #if defined(UNOPTIMIZED_BUILD)
   constexpr uint8_t percentExpected = 60;
 #else
-  constexpr uint8_t percentExpected = 40;
+  constexpr uint8_t percentExpected = 35;
 #endif 
   performance_test(12, dividendGen, divisorGen, percentExpected);
 }
@@ -99,7 +99,7 @@ static void test_fast_div_perf_u32_u16_worst_case(void)
   static constexpr index_range_generator<uint32_t> dividendGen(divisorGen.rangeMax()*2ULL, UINT32_MAX, divisorGen.num_steps());
 
 #if defined(UNOPTIMIZED_BUILD)
-  constexpr uint8_t percentExpected = 110;
+  constexpr uint8_t percentExpected = 105;
 #else
   constexpr uint8_t percentExpected = 105;
 #endif 
@@ -112,9 +112,17 @@ static void test_fast_div_perf_u16_u16(void)
   static constexpr index_range_generator<uint16_t> dividendGen(divisorGen.rangeMax()+1U, UINT16_MAX, divisorGen.num_steps());
 
 #if defined(UNOPTIMIZED_BUILD)
-  constexpr uint8_t percentExpected = 65;
+#if defined(AFD_SMALL_TEXT)
+  constexpr uint8_t percentExpected = 70;
+#else
+  constexpr uint8_t percentExpected = 70;
+#endif
+#else
+#if defined(AFD_SMALL_TEXT)
+  constexpr uint8_t percentExpected = 50;
 #else
   constexpr uint8_t percentExpected = 45;
+#endif
 #endif 
   performance_test(3, dividendGen, divisorGen, percentExpected);
 }
@@ -127,7 +135,7 @@ static void test_fast_div_perf_u32_u32(void)
 #if defined(UNOPTIMIZED_BUILD)
   constexpr uint8_t percentExpected = 80;
 #else
-  constexpr uint8_t percentExpected = 75;
+  constexpr uint8_t percentExpected = 65;
 #endif 
   performance_test(1, dividendGen, divisorGen, percentExpected);
 }
@@ -151,9 +159,17 @@ static void test_fast_div_perf_s32_s16_worst_case(void)
   static constexpr index_range_generator<int32_t> dividendGen(divisorGen.rangeMax()*3LL, INT32_MAX/55LL, divisorGen.num_steps());
 
 #if defined(UNOPTIMIZED_BUILD)
+#if defined(AFD_SMALL_TEXT)
+  constexpr uint8_t percentExpected = 65;
+#else
   constexpr uint8_t percentExpected = 60;
+#endif
+#else
+#if defined(AFD_SMALL_TEXT)
+  constexpr uint8_t percentExpected = 50;
 #else
   constexpr uint8_t percentExpected = 45;
+#endif
 #endif 
   performance_test(32, dividendGen, divisorGen, percentExpected);
 }
